임시 메타정보 기입

+++ src/main/java/com/sbs/tutorial/app1/boudedContext/app/fileUpload/entity/GenFile.java
@@ -14,7 +14,7 @@ import lombok.experimental.SuperBuilder;
 @ToString(callSuper = true)
 public class GenFile extends BaseEntity {
   private String relTypeCode;
-  private int relId;
+  private Long relId;
   private String typeCode;
   private String type2Code;
   private String fileExtTypeCode;
